build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(lwadvapi: r39716)

+++ lwadvapi/unthreaded/memory/lwstr.c
@@ -360,7 +360,7 @@ LwStrToUpper(
         return;
 
     while (*pszString != '\0') {
-        *pszString = toupper(*pszString);
+        *pszString = toupper((int)*pszString);
         pszString++;
     }
 }
@@ -376,7 +376,7 @@ LwStrnToUpper(
        DWORD iCh = 0;
 
        while (*pszString != '\0' && (iCh++ < dwLen)) {
-           *pszString = toupper(*pszString);
+           *pszString = toupper((int)*pszString);
            pszString++;
        }
     }
@@ -391,7 +391,7 @@ LwStrToLower(
         return;
 
     while (*pszString != '\0') {
-        *pszString = tolower(*pszString);
+        *pszString = tolower((int)*pszString);
         pszString++;
     }
 }
@@ -408,7 +408,7 @@ LwStrnToLower(
         DWORD iCh = 0;
 
         while (*pszString != '\0' && (iCh++ < dwLen)) {
-            *pszString = tolower(*pszString);
+            *pszString = tolower((int)*pszString);
             pszString++;
         }
     }
